requirements.txt
flask==3.1.1
waitress==3.0.2
pathvalidate==3.3.1
requests==2.32.4
pycryptodome==3.23.0
click==8.2.1
prompt-toolkit==3.0.51
psycopg[binary]==3.2.9
psycopg-pool==3.2.6
